+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0]
+### Added
+- Xls reader produces numeric values conforming to Xlsx reader.
+- Xls reader correctly estimates number of lines in sheet.
+- Xls reader tests.
+
 ## [0.4.0] - 2022-06-08
 ### Added
 - Lines returned by LineReaders changed from List of Objects to Map of Strings